Company detail and update added
