+++ web-react/src/components/Company/CompanyDetails.css
@@ -28,7 +28,7 @@
   padding: 0;
 }
 
-.MuiPaper-root.domainDeatils {
+.MuiPaper-root.companyDetails {
   margin-top: 3px;
 }
