+++ web-react/src/components/List.js
@@ -56,8 +56,9 @@ function List({ data,
                   return (
                     <TableRow key={i}>
                       <TableCell component="th" scope="row">
-                        <Link to={{ pathname: `/details${title}`, state: { name: n.name } }}  >
+                        <Link to={{ pathname: `/details${title}`, state: { id: n.id } }}  >
                           {n.name}
+
                         </Link>
                       </TableCell>
                       <TableCell>{n.description}</TableCell>
